requirements.txt
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
chardet==3.0.4
charset-normalizer==3.3.2
contourpy==1.3.0
cycler==0.12.1
deep-translator==1.11.4
et-xmlfile==1.1.0
Faker==28.4.1
fonttools==4.53.1
google-api-core==2.19.2
google-auth==2.34.0
google-cloud-core==2.4.1
google-cloud-translate==3.16.0
googleapis-common-protos==1.65.0
googletrans==4.0.0rc1
grpc-google-iam-v1==0.13.1
grpcio==1.66.1
grpcio-status==1.66.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.9.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
kiwisolver==1.4.7
langdetect==1.0.9
matplotlib==3.9.2
numpy==2.1.1
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
tzdata==2024.1
urllib3==2.2.2
